FTS: Fuzzy matching, boosted queries, sloppy phrase matching, code tidy up, add API for specifying Analysis type (MOB-571)
git-svn-id: https://svn.alfresco.com/repos/alfresco-enterprise/alfresco/HEAD/root@14173 c4b6b30b-aa2e-2d43-bbcb-ca4b014f7261

+++ source/java/org/alfresco/repo/search/impl/querymodel/impl/functions/FTSPhrase.java
@@ -56,7 +56,7 @@ public class FTSPhrase extends BaseFunction
         args = new LinkedHashMap<String, ArgumentDefinition>();
         args.put(ARG_PHRASE, new BaseArgumentDefinition(Multiplicity.SINGLE_VALUED, ARG_PHRASE, DataTypeDefinition.ANY, true));
         args.put(ARG_PROPERTY, new BaseArgumentDefinition(Multiplicity.SINGLE_VALUED, ARG_PROPERTY, DataTypeDefinition.ANY, false));
-        args.put(ARG_SLOP, new BaseArgumentDefinition(Multiplicity.SINGLE_VALUED, ARG_SLOP, DataTypeDefinition.FLOAT, false));
+        args.put(ARG_SLOP, new BaseArgumentDefinition(Multiplicity.SINGLE_VALUED, ARG_SLOP, DataTypeDefinition.INT, false));
     }
 
     /**
